+++ src/realm/object-store/c_api/util.hpp
@@ -2,6 +2,7 @@
 #define REALM_OBJECT_STORE_C_API_UTIL_HPP
 
 #include <realm/object-store/c_api/types.hpp>
+#include <realm/util/functional.hpp>
 
 namespace realm::c_api {
 
@@ -115,6 +116,33 @@ struct FreeUserdata {
 };
 
 using UserdataPtr = std::unique_ptr<void, FreeUserdata>;
+
+template <typename... Args>
+class CallbackRegistry {
+public:
+    uint64_t add(util::UniqueFunction<void(Args...)>&& callback)
+    {
+        uint64_t token = m_next_token++;
+        m_callbacks.emplace(token, std::move(callback));
+        return token;
+    }
+
+    void remove(uint64_t token)
+    {
+        m_callbacks.erase(token);
+    }
+
+    void invoke(Args... args)
+    {
+        for (auto& callback : m_callbacks) {
+            callback.second(std::forward<Args>(args)...);
+        }
+    }
+
+private:
+    std::map<uint64_t, util::UniqueFunction<void(Args...)>> m_callbacks;
+    uint64_t m_next_token;
+};
 } // namespace realm::c_api
 
 #endif // REALM_OBJECT_STORE_C_API_UTIL_HPP